Having problems with Gaussian
First, rounding errors. annoying, but not too severe. second. Image brightness values are diverging rather than converging... not too sure why, but i think it may just be the kernel size is not large enough, so areas that are whit are becoming more white, which end up making a second pass create brighter areas and darker areas... total brightness average preserved, but local averages tending to the extremes.... no fun... will try larger kernel later, just to see if it works. If not i might just find a library, because this is not good progress :'(