Added 5x5 gaussian kernel
Noise is still there?!?!?! it is visibly significantly less... maybe i need a particularly larger kernel because the image is large? i honestly dont know. will try 7 and 9, but at least the speed up by separating them will be larger, even if its slower overall...

Also added a skip for the first 5 frames, because they are always junk while the camera is turning on and focusing.