Added a few different gaussian kernels
The noise is much more reduced in the larger sigma, however it is now starting to form pockets of anomolous brightness values. Its much better than it was, so i think i need to use a larger kernel (7 or 9 maybe?) with a large standard deviation (sigma=0.5*radius?). I didnt want to really blur it by that much, but it seems that my camera is very noisy in dark situations...