Added 7x7 gaussian with sd3.5
Larger gaussian with wider spread average seems to have removed much more of what im now calling (for some reason) "cancer spots", because they kind of look like weird areas that should behave like they do, and somehow want to merge together to form a larger spot if given the chance ;P im starting to think that it is only when there is a significant brightness change with a lot of natural noise around it, so a peice of paper should fare fine. a larger kernel might reduce it even further, and maybe ill do a 9x9 later down the line just to be safe, but the blur is starting to take some time pre-separated, so ill leave it at "good enough" for now, and move on to the edge detection part. Again.