Figured out gaussian issue
I believe the "cancer-spots" are being created in areas where the brightness differential is particularly high, eg bright white on dark black, with no gradient between them. This will cause the kernel above the border to output a very bright value, as the majority of the pixels are bright, and 3/4 of the second largest kernal grid is bright, and the kernel below the border will be the exact opposite. The only way to combat this is to use a larger kernel with a higher standard deviation. Which takes longer. A lot longer. Oh well...